requirements.txt
Babel==2.10.3
certifi==2021.10.8
cycler==0.11.0
distlib==0.3.5
dnspython==2.2.0
et-xmlfile==1.1.0
filelock==3.7.1
fonttools==4.33.3
fpdf==1.7.2
kiwisolver==1.4.2
matplotlib==3.5.2
mysql-connector-python==8.0.29
numpy==1.22.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.0
Pillow==9.1.1
platformdirs==2.5.2
PollyReports==1.8
prettytable==3.3.0
protobuf==4.21.1
pwinput==1.0.2
pymongo==4.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.6
six==1.16.0
tabulate==0.8.9
tkcalendar==1.6.1
virtualenv==20.15.1
wcwidth==0.2.5
XlsxWriter==3.0.2
